Further refactoring data_io
