+++ openretina/data_io/artificial_stimuli.py
