+++ openretina/data_io/hoefling_2024/dataloaders.py
@@ -1,8 +1,15 @@
-from typing import Any
+from typing import Any, Optional
 
 import numpy as np
 import torch
 from torch.utils.data import default_collate
+from tqdm.auto import tqdm
+
+from openretina.data_io.artificial_stimuli import load_chirp, load_moving_bar
+from openretina.data_io.hoefling_2024.constants import CLIP_LENGTH, NUM_CLIPS, NUM_VAL_CLIPS
+from openretina.data_io.hoefling_2024.responses import NeuronData
+from openretina.data_io.hoefling_2024.stimuli import gen_start_indices, get_all_movie_combinations
+from openretina.data_io.movie_dataloader import MoviesTrainTestSplit, get_movie_dataloader
 
 
 def get_dims_for_loader_dict(dataloaders: dict[str, dict[str, Any]]) -> dict[str, dict[str, tuple[int, ...]] | tuple]:
@@ -118,3 +125,222 @@ def extract_data_info_from_dataloaders(
         data_info[session_key]["roi_coords"] = np.array(dataloaders[session_key].dataset.roi_coords)  # type: ignore
 
     return data_info
+
+
+def natmov_dataloaders_v2(
+    neuron_data_dictionary: dict[str, Any],
+    movies_dictionary: MoviesTrainTestSplit,
+    train_chunk_size: int = 50,
+    batch_size: int = 32,
+    num_clips: int = NUM_CLIPS,
+    clip_length: int = CLIP_LENGTH,
+    num_val_clips: int = NUM_VAL_CLIPS,
+):
+    assert all(field in next(iter(neuron_data_dictionary.values())) for field in ["responses_final", "stim_id"]), (
+        "Check the neuron data dictionary sub-dictionaries for the minimal"
+        " required fields: 'responses_final' and 'stim_id'."
+    )
+    assert (
+        next(iter(neuron_data_dictionary.values()))["stim_id"] == 5
+    ), "This function only supports natural movie stimuli."
+
+    # Draw validation clips based on the random seed
+    val_clip_idx = list(np.random.choice(num_clips, num_val_clips, replace=False))
+
+    clip_chunk_sizes = {
+        "train": train_chunk_size,
+        "validation": clip_length,
+        "test": movies_dictionary.test.shape[1],
+    }
+    dataloaders: dict[str, Any] = {"train": {}, "validation": {}, "test": {}}
+
+    # Get the random sequences of movies presentations for each session if available
+    if movies_dictionary.random_sequences is None:
+        movie_length = movies_dictionary.train.shape[1]
+        random_sequences = np.arange(0, movie_length // clip_length)[:, np.newaxis]
+    else:
+        random_sequences = movies_dictionary.random_sequences
+
+    movies = get_all_movie_combinations(
+        movies_dictionary.train,
+        movies_dictionary.test,
+        random_sequences,
+        val_clip_idx=val_clip_idx,
+        num_clips=num_clips,
+        clip_length=clip_length,
+    )
+
+    start_indices = gen_start_indices(random_sequences, val_clip_idx, clip_length, train_chunk_size, num_clips)
+
+    for session_key, session_data in tqdm(neuron_data_dictionary.items(), desc="Creating movie dataloaders"):
+        neuron_data = NeuronData(
+            **session_data,
+            random_sequences=random_sequences,  # Used together with the validation index to
+            # get the validation response in the corresponding dict
+            val_clip_idx=val_clip_idx,
+            num_clips=num_clips,
+            clip_length=clip_length,
+        )
+        _eye = neuron_data.eye
+
+        if session_key == "session_2_ventral2_20200626":
+            # session incorrectly labeled as left
+            _eye = "right"
+        # if neuron_data.responses_train.shape[-1] == 0:
+        #     print("skipped: {}".format(session_key))
+        #     break
+        for fold in ["train", "validation", "test"]:
+            dataloaders[fold][session_key] = get_movie_dataloader(
+                movies=movies[_eye][fold],
+                responses=neuron_data.response_dict[fold],
+                roi_ids=neuron_data.roi_ids,
+                roi_coords=neuron_data.roi_coords,
+                group_assignment=neuron_data.group_assignment,
+                scan_sequence_idx=neuron_data.scan_sequence_idx,
+                split=fold,
+                chunk_size=clip_chunk_sizes[fold],
+                start_indices=start_indices[fold],
+                batch_size=batch_size,
+                scene_length=clip_length,
+            )
+
+    return dataloaders
+
+
+def get_chirp_dataloaders(
+    neuron_data_dictionary,
+    train_chunk_size: Optional[int] = None,
+    batch_size: int = 32,
+):
+    assert isinstance(
+        neuron_data_dictionary, dict
+    ), "neuron_data_dictionary should be a dictionary of sessions and their corresponding neuron data."
+    assert all(
+        field in next(iter(neuron_data_dictionary.values()))
+        for field in ["responses_final", "stim_id", "chirp_trigger_times"]
+    ), (
+        "Check the neuron data dictionary sub-dictionaries for the minimal required fields: "
+        "'responses_final', 'stim_id' and 'chirp_trigger_times'."
+    )
+
+    assert next(iter(neuron_data_dictionary.values()))["stim_id"] == 1, "This function only supports chirp stimuli."
+
+    dataloaders: dict[str, Any] = {"train": {}}
+
+    chirp_triggers = next(iter(neuron_data_dictionary.values()))["chirp_trigger_times"][0]
+    # 2 triggers per chirp presentation
+    num_chirps = len(chirp_triggers) // 2
+
+    # Get it into chan, time, height, width
+    chirp_stimulus = torch.tensor(load_chirp(), dtype=torch.float32).permute(3, 0, 1, 2)
+
+    chirp_stimulus = chirp_stimulus.repeat(1, num_chirps, 1, 1)
+
+    # Use full chirp for training if no chunk size is provided
+    clip_chunk_sizes = {
+        "train": train_chunk_size if train_chunk_size is not None else chirp_stimulus.shape[1] // num_chirps,
+    }
+
+    # 5 chirp presentations
+    start_indices = np.arange(0, chirp_stimulus.shape[1] - 1, chirp_stimulus.shape[1] // num_chirps).tolist()
+
+    for session_key, session_data in tqdm(neuron_data_dictionary.items(), desc="Creating chirp dataloaders"):
+        neuron_data = NeuronData(
+            **session_data,
+            random_sequences=None,
+            val_clip_idx=None,
+            num_clips=None,
+            clip_length=None,
+        )
+
+        session_key += "_chirp"
+
+        dataloader = get_movie_dataloader(
+            movies=chirp_stimulus if neuron_data.eye == "right" else torch.flip(chirp_stimulus, [-1]),
+            responses=neuron_data.response_dict["train"],
+            roi_ids=neuron_data.roi_ids,
+            roi_coords=neuron_data.roi_coords,
+            group_assignment=neuron_data.group_assignment,
+            scan_sequence_idx=neuron_data.scan_sequence_idx,
+            split="train",
+            chunk_size=clip_chunk_sizes["train"],
+            start_indices=start_indices,
+            batch_size=batch_size,
+            scene_length=chirp_stimulus.shape[1] // num_chirps,
+            drop_last=False,
+        )
+        if dataloader is not None:
+            dataloaders["train"][session_key] = dataloader
+        else:
+            print(f"Ignoring session {session_key} for stimulus chirp")
+
+    return dataloaders
+
+
+def get_mb_dataloaders(
+    neuron_data_dictionary,
+    train_chunk_size: Optional[int] = None,
+    batch_size: int = 32,
+):
+    assert isinstance(
+        neuron_data_dictionary, dict
+    ), "neuron_data_dictionary should be a dictionary of sessions and their corresponding neuron data."
+    assert all(
+        field in next(iter(neuron_data_dictionary.values()))
+        for field in ["responses_final", "stim_id", "mb_trigger_times"]
+    ), (
+        "Check the neuron data dictionary sub-dictionaries for the minimal required fields: "
+        "'responses_final', 'stim_id' and 'mb_trigger_times'."
+    )
+
+    assert (
+        next(iter(neuron_data_dictionary.values()))["stim_id"] == 2
+    ), "This function only supports moving bar stimuli."
+
+    dataloaders: dict[str, Any] = {"train": {}}
+
+    mb_triggers = next(iter(neuron_data_dictionary.values()))["mb_trigger_times"][0]
+    num_repeats = len(mb_triggers) // 8
+
+    # Get it into chan, time, height, width
+    mb_stimulus = torch.tensor(load_moving_bar(), dtype=torch.float32).permute(3, 0, 1, 2)
+
+    mb_stimulus = mb_stimulus.repeat(1, num_repeats, 1, 1)
+
+    # 8 directions
+    total_num_mbs = 8 * num_repeats
+
+    # Default to each mb for training if no chunk size provided.
+    clip_chunk_sizes = {
+        "train": train_chunk_size if train_chunk_size is not None else mb_stimulus.shape[1] // total_num_mbs,
+    }
+
+    start_indices = np.arange(0, mb_stimulus.shape[1] - 1, step=mb_stimulus.shape[1] // total_num_mbs).tolist()
+
+    for session_key, session_data in tqdm(neuron_data_dictionary.items(), desc="Creating moving bars dataloaders"):
+        neuron_data = NeuronData(
+            **session_data,
+            random_sequences=None,
+            val_clip_idx=None,
+            num_clips=None,
+            clip_length=None,
+        )
+
+        session_key += "_mb"
+
+        dataloaders["train"][session_key] = get_movie_dataloader(
+            movies=mb_stimulus if neuron_data.eye == "right" else torch.flip(mb_stimulus, [-1]),
+            responses=neuron_data.response_dict["train"],
+            roi_ids=neuron_data.roi_ids,
+            roi_coords=neuron_data.roi_coords,
+            group_assignment=neuron_data.group_assignment,
+            scan_sequence_idx=neuron_data.scan_sequence_idx,
+            split="train",
+            chunk_size=clip_chunk_sizes["train"],
+            start_indices=start_indices,
+            batch_size=batch_size,
+            scene_length=mb_stimulus.shape[1] // total_num_mbs,
+            drop_last=False,
+        )
+
+    return dataloaders
