+++ scripts/train_core_readout.py
@@ -10,8 +10,8 @@ from lightning.pytorch.callbacks import ModelCheckpoint
 from omegaconf import DictConfig, OmegaConf
 
 from openretina.data_io.cyclers import LongCycler
-from openretina.data_io.hoefling_2024.data_io import natmov_dataloaders_v2
-from openretina.data_io.hoefling_2024.neuron_data_io import filter_responses, make_final_responses
+from openretina.data_io.hoefling_2024.dataloaders import natmov_dataloaders_v2
+from openretina.data_io.hoefling_2024.responses import filter_responses, make_final_responses
 from openretina.data_io.movie_dataloader import MoviesTrainTestSplit
 from openretina.models.core_readout import CoreReadout
 from openretina.utils.h5_handling import load_h5_into_dict
